Relocate Community/JSON-LD drift guard to its own focused test file
The guard verifies that Community.vocabBase matches the JSON-LD @context
fixture's `esea` prefix. It is a config-coherence test, not URL helper
behavior. Living inside the expandFacets/compactFacets describe block
implied it tested those helpers (it does not) and dragged fs/path imports
into the URL helper tests.

Move to tests/services/communities.test.ts. The guard plus its comment
explaining what it protects against now stand alone.

No behavior change; 492 tests still green.
